Firebase connection with the app
The access was created in Firebase for the app, to be able to access Google, through the plugin @react-native-community/google-signin.

Two buttons were placed, one for logging in and the other for logging out.

When you log in or when you log out, it is displayed in the 'npx react-native start' terminal or in the debug console.

A library was also added to access the Environment Variables and, in that way, hide the sensitive information of the app, such as the web client ID.